+++ postgres.go
@@ -4,20 +4,85 @@ import (
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/postgres"
 	"reflect"
+	"fmt"
+	"time"
+	"log"
 )
 
 type PostgresCoreRepo struct {
 	DB *gorm.DB
 }
 
-func NewPostgesCoreRepo(connectionString string) *PostgresCoreRepo {
-	db, err := gorm.Open("postgres", connectionString)
+type PostgresCoreRepoOptions struct {
+	Host            string
+	User            string
+	DB              string
+	SSLMode         string
+	Password        string
+	Log             bool
+	MaxIdleConns    int
+	MaxOpensConns   int
+	CoonMaxLifetime time.Duration
+}
+
+func NewPostgresCoreRepoOptions() PostgresCoreRepoOptions {
+	options := PostgresCoreRepoOptions{}
+	options.Log = false
+	options.MaxIdleConns = -1
+	options.MaxOpensConns = -1
+	options.CoonMaxLifetime = -1
+	return options
+}
+
+func NewPostgresCoreRepo(options PostgresCoreRepoOptions) *PostgresCoreRepo {
+
+	if options.Host == "" {
+		log.Fatalln("options.Host is missing.")
+	}
+
+	if options.User == "" {
+		log.Fatalln("options.User is missing.")
+	}
+
+	if options.DB == "" {
+		log.Fatalln("options.DB is missing.")
+	}
+
+	if options.SSLMode == "" {
+		log.Fatalln("options.SSLMode is missing.")
+	}
+
+	if options.Password == "" {
+		log.Fatalln("options.Password is missing.")
+	}
+
+	connection := fmt.Sprintf("host=%s user=%s dbname=%s sslmode=%s password=%s",
+		options.Host,
+		options.User,
+		options.DB,
+		options.SSLMode,
+		options.Password,
+	)
+
+	db, err := gorm.Open("postgres", connection)
 
 	if err != nil {
 		panic("failed to connect database")
 	}
 
-	db.LogMode(true)
+	db.LogMode(options.Log)
+
+	if options.MaxIdleConns != -1 {
+		db.DB().SetMaxIdleConns(options.MaxIdleConns)
+	}
+
+	if options.MaxOpensConns != -1 {
+		db.DB().SetMaxOpenConns(options.MaxOpensConns)
+	}
+
+	if options.CoonMaxLifetime != -1 {
+		db.DB().SetConnMaxLifetime(options.CoonMaxLifetime)
+	}
 
 	return &PostgresCoreRepo{
 		DB: db,
@@ -32,7 +97,7 @@ func (rp *PostgresCoreRepo) Migrations(models []interface{}) {
 	}
 }
 
-func setPreload (db *gorm.DB, preload []string) *gorm.DB {
+func setPreload(db *gorm.DB, preload []string) *gorm.DB {
 	if preload != nil {
 		for _, elm := range preload {
 			db = db.Preload(elm)
@@ -105,30 +170,14 @@ func (rp *PostgresCoreRepo) GetWhereMultiple(model interface{}, preload []string
 }
 
 func (rp *PostgresCoreRepo) Update(id string, model interface{}) error {
-	if err := rp.DB.Save(model).Error; err != nil {
-		return err
-	}
-
-	return nil
+	return rp.DB.Save(model).Error
 }
 
 // TODO Remove associations
 func (rp *PostgresCoreRepo) Delete(id string, model interface{}) error {
-	if err := rp.DB.Where("id = ?", id).Unscoped().Delete(model).Error; err != nil {
-		return err
-	}
-
-	return nil
+	return rp.DB.Where("id = ?", id).Unscoped().Delete(model).Error
 }
 
 func (rp *PostgresCoreRepo) Create(model interface{}) error {
-	tx := rp.DB.Begin()
-
-	if err := rp.DB.Create(model).Error; err != nil {
-		tx.Rollback()
-		return err
-	}
-
-	tx.Commit()
-	return nil
+	return rp.DB.Create(model).Error
 }