+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 ==================
 
   * Added GetWhereMultiple
+
+0.5.0 / 2017-07-05
+==================
+
+  * Add PostgresCoreRepoOptions and NewPostgresCoreRepoOptions
+  * Options validations on NewPostgresCoreRepo
+  * Remove transaction from Create